+++ include/cobalt/cobalt.h
@@ -69,6 +69,9 @@ CO_API int co_get_task(struct cobalt *co, uint32_t id, struct cobalt_query **q);
 CO_API int co_get_board(struct cobalt *co, const char *board,
 		struct cobalt_query **q);
 
+/* get all tasks from the database */
+CO_API int co_get_all(struct cobalt *co, struct cobalt_query **q);
+
 /* reading the results of a query */
 CO_API struct cobalt_query *co_query_getnext(struct cobalt *co,
 		struct cobalt_query *q);